% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{consortr}
\alias{consortr}
\title{Shiny app for generating consort flow diagrams and other types of flow diagrams}
\usage{
consortr()
}
\value{
none
}
\description{
A consort diagram graphically depicts the passage of participants through a randomized clinical trial.
This app can be used to easily create consort diagrams, and to visualize any other process where criteria are applied
in succession to a dataset and it is of interest to know how many rows of the dataset remain after the application of each criterion.
}
\references{
Moher, Schulz and Altman (2001)
The CONSORT statement: revised recommendations for improving the quality of reports of parallel-group randomised trials.
\emph{Lancet} \bold{357}, 1191-94.
}
