% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{with_dir}
\alias{with_dir}
\title{Run R code in temporary working directory.}
\usage{
with_dir(dir, expr)
}
\arguments{
\item{dir}{Scalar string giving the desired temporary working directory.}

\item{expr}{Expression to run with the temporary working directory in effect.}
}
\value{
Invisible NULL
}
\description{
Sometimes one might want to run R code with a temporary working directory in
effect.  This function will do that
}
\examples{
with_dir(tempdir(), cat("Hello world!", file=tempfile(fileext=".txt")))
}
